生意人文章内容完善.推荐.轮播.评分.专题管理及接口+生意人用户.角色.菜单管理+平台用户审核(未完成)

+++ portal-webapp/portal-admin/src/main/java/com/yonghui/portal/service/app/AppUserRolesService.java
@@ -9,7 +9,7 @@ import java.util.List;
  */
 public interface AppUserRolesService {
 
-    void saveOrUpdate(Long userId, List<Integer> roleIdList);
+    void saveOrUpdate(Integer userId, List<Integer> roleIdList);
 
     /**
      * 根据用户ID,获取角色ID列表
